==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data necessary
# Authors: Sami El Sabri, Liban Timir
# Date: 10 February 2023
# Contact: sami.elsabri@mail.utoronto.ca, liban.timir@mail.utoronto.ca
# License: MIT
# Pre-requisites: Download 1.3GB large file titled 'WVS TimeSeries 1981 2022 Csv v4 0.zip' from 
# https://www.worldvaluessurvey.org/WVSDocumentationWVL.jsp


#### Workspace setup ####
library(tidyverse)

#### Download data ####

# wvs_data <- read_csv("inputs/data/study_3/WVS_TimeSeries_4_0.csv")
# This file is too big to add to this folder and to github. 
# Follow above instructions and add the file to the specified filepath
