==> scripts/06-replication-table2.R <==
#### Preamble ####
# Purpose: Replicate Table 2 of Feldman et al. (2017)
# Authors: Sami El Sabri, Liban Timir
# Date: 10 February 2023
# Contact: sami.elsabri@mail.utoronto.ca
# License: MIT

library(tidyverse)
library(ggplot2)
library(psych)
library(knitr)
library(haven)

satisfaction_data_usa <- read_csv("inputs/data/study_2/FW satisfaction-Study 2-data.csv")


satisfaction_data_usa_summary <- satisfaction_data_usa %>% 
  summarize(
    mean_js1 = round(mean(jobsat1, na.rm = TRUE), 2),
    sd_js1 = round(sd(jobsat1, na.rm = TRUE), 2),
    mean_js2 = round(mean(jobsat2, na.rm = TRUE), 2),
    sd_js2 = round(sd(jobsat2, na.rm = TRUE), 2),
    mean_freewill = round(mean(FWDfwagency, na.rm = TRUE), 2),
    sd_freewill = round(sd(FWDfwagency, na.rm = TRUE), 2),
    mean_jobaut = round(mean(jobaut, na.rm = TRUE), 2),
    sd_jobaut = round(sd(jobaut, na.rm = TRUE), 2),
    mean_jobaut2 = round(mean(jobaut2, na.rm = TRUE), 2),
    sd_jobaut2 = round(sd(jobaut2, na.rm = TRUE), 2),
    mean_locus = round(mean(locus, na.rm = TRUE), 2),
    sd_locus = round(sd(locus, na.rm = TRUE), 2),
    mean_selfcontrol = round(mean(selfcontrol, na.rm = TRUE), 2),
    sd_selfcontrol = round(sd(selfcontrol, na.rm = TRUE), 2),
    mean_selfeff = round(mean(selfeff, na.rm = TRUE), 2),
    sd_selfeff = round(sd(selfeff, na.rm = TRUE), 2),
    mean_selfest = round(mean(selfest, na.rm = TRUE), 2),
    sd_selfest = round(sd(selfest, na.rm = TRUE), 2),
    mean_implicit = round(mean(ess_kind, na.rm = TRUE), 2),
    sd_implicit = round(sd(ess_kind, na.rm = TRUE), 2)
  )

calculate_correlations <- function(data, target_var) {
  # Identify all other variables in the dataset
  other_vars <- setdiff(names(data), target_var)
  
  # Use map_dfr to iterate over each variable and calculate correlation with target_var
  correlations <- map_dfr(other_vars, function(var) {
    data_subset <- data %>% 
      select(all_of(c(target_var, var))) %>% 
      na.omit()  # Ensure pairwise complete observation
    
    correlation <- cor(data_subset[[1]], data_subset[[2]], method = "pearson")
    
    # Return a data frame with the variable name and its correlation with target_var
    tibble(Variable = var, Correlation = correlation)
  })
  
  return(correlations)
}

# Construct Correlations
satisfaction_data_usa_rel <- satisfaction_data_usa %>% select("jobsat1", "jobsat2", "FWDfwagency", "jobaut",
                                                              "jobaut2", "locus", "ess_kind", "selfest", "selfeff", "selfcontrol")
correlations_jobsat1 <- calculate_correlations(satisfaction_data_usa_rel, "jobsat1")
correlations_jobsat1 <- round(correlations_jobsat1[1:9, 2],2)
correlations_jobsat2 <-  calculate_correlations(satisfaction_data_usa_rel, "jobsat2")
correlations_jobsat2 <- round(correlations_jobsat2[2:9,2],2)
correlations_fw <-  calculate_correlations(satisfaction_data_usa_rel, "FWDfwagency")
correlations_fw <- round(correlations_fw[3:9,2],2)
correlations_jobaut1 <-  calculate_correlations(satisfaction_data_usa_rel, "jobaut")
correlations_jobaut1 <- round(correlations_jobaut1[4:9,2],2)
correlations_jobaut2 <- calculate_correlations(satisfaction_data_usa_rel, "jobaut2")
correlations_jobaut2 <- round(correlations_jobaut2[5:9,2],2)

## Construct Alpha Values
js1_items <- satisfaction_data_usa[, c("jobsat_1", "jobsat_2", "jobsat_3", "jobsat_4", "jobsat_5")]
alpha_js1 <- alpha(js1_items, check.keys = TRUE)
alpha_js1 <- alpha_js1$total$raw_alpha
alpha_js1 <- round(alpha_js1,2)

js2_items <- satisfaction_data_usa[, c("jobsat2_1", "jobsat2_2", "jobsat2_3", "jobsat2_4", "jobsat2_5")]
alpha_js2 <- alpha(js2_items, check.keys = TRUE)
alpha_js2 <- alpha_js2$total$raw_alpha
alpha_js2 <- round(alpha_js2,2)

freewill_items <- satisfaction_data_usa[, c("fw_1", "fw_2", "fw_3", "fw_4", "fw_5", "fw_6", "fw_7", "fw_8", "fw_9")]
alpha_freewill <- alpha(freewill_items)
alpha_freewill <- alpha_freewill$total$raw_alpha
alpha_freewill <- round(alpha_freewill,2)

jaut1_items <- satisfaction_data_usa[, c("jobaut_3", "jobaut_4", "jobaut_5")]
alpha_jaut1 <- alpha(jaut1_items)
alpha_jaut1 <- alpha_jaut1$total$raw_alpha
alpha_jaut1 <- round(alpha_jaut1,2)

jaut2_items <- satisfaction_data_usa[, c("jobaut2_3", "jobaut2_4", "jobaut2_5")]
alpha_jaut2 <- alpha(jaut2_items)
alpha_jaut2 <- alpha_jaut2$total$raw_alpha
alpha_jaut2 <- round(alpha_jaut2,2)

locus_items <- satisfaction_data_usa[, c("lc1", "lc2", "lc3", "lc4", "lc5", "lc6", "lc7", "lc8", "lc9", "lc10", "lc11", "lc12", "lc13")]
alpha_locus <- alpha(locus_items, check.keys=TRUE)
alpha_locus <- alpha_locus$total$raw_alpha
alpha_locus <- round(alpha_locus,2)

self_eff_items <- satisfaction_data_usa[, c("selfeff_1", "selfeff_2", "selfeff_3")]
alpha_self_eff <- alpha(self_eff_items)
alpha_self_eff <- alpha_self_eff$total$raw_alpha
alpha_self_eff <- round(alpha_self_eff,2)

self_est_items <- satisfaction_data_usa[, c("selfest_1", "selfest_2", "selfest_3", "selfest_4", "selfest_5", "selfest_6", "selfest_7", "selfest_8", "selfest_9", "selfest_10")]
alpha_self_est <- alpha(self_est_items, check.keys=TRUE)
alpha_self_est <- alpha_self_est$total$raw_alpha
alpha_self_est <- round(alpha_self_est,2)

self_con_items <- satisfaction_data_usa[, c("selfcon_1", "selfcon_2", "selfcon_3", "selfcon_4", "selfcon_5", "selfcon_6", "selfcon_7", "selfcon_8", "selfcon_9", "selfcon_10", "selfcon_11", "selfcon_12", "selfcon_13")]
alpha_self_con <- alpha(self_con_items, check.keys=TRUE)
alpha_self_con <- alpha_self_con$total$raw_alpha
alpha_self_con <- round(alpha_self_con,2)

implicit_items <- satisfaction_data_usa[, c("lay_1", "lay_2", "lay_3", "lay_4", "lay_5r", "lay_6r", "lay_7r", "lay_8r")]
alpha_implicit <- alpha(implicit_items, check.keys=TRUE)
alpha_implicit <- alpha_implicit$total$raw_alpha
alpha_implicit <- round(alpha_implicit,2)




## Construct Final Table
final_table <- data.frame(
  Variable = c("Job Satisfaction (T1)", "Job Satisfaction (T2)", "Belief in Free Will (T1)",
               "Job Autonomy (T1)", "Job Autonomy (T2)", "Locus of Control", "Implicit Beliefs",
               "Self-Esteem", "Self-Efficacy", "Self-Control"),
  "Reliability Coefficient" = c(alpha_js1, alpha_js2, alpha_freewill, alpha_jaut1, alpha_jaut2,
                                NA, NA, NA, NA, NA),
  Mean = c(
           satisfaction_data_usa_summary$mean_js1, 
           satisfaction_data_usa_summary$mean_js2,
           satisfaction_data_usa_summary$mean_freewill,
           satisfaction_data_usa_summary$mean_jobaut,
           satisfaction_data_usa_summary$mean_jobaut2,
           satisfaction_data_usa_summary$mean_locus,
           satisfaction_data_usa_summary$mean_implicit,
           satisfaction_data_usa_summary$mean_selfest,
           satisfaction_data_usa_summary$mean_selfeff,
           satisfaction_data_usa_summary$mean_selfcontrol),
  SD = c(satisfaction_data_usa_summary$sd_js1, 
         satisfaction_data_usa_summary$sd_js2,
         satisfaction_data_usa_summary$sd_freewill, 
         satisfaction_data_usa_summary$sd_jobaut,
         satisfaction_data_usa_summary$sd_jobaut2,
         satisfaction_data_usa_summary$sd_locus,
         satisfaction_data_usa_summary$sd_implicit,
         satisfaction_data_usa_summary$sd_selfest,
         satisfaction_data_usa_summary$sd_selfeff,
         satisfaction_data_usa_summary$sd_selfcontrol),
  "Job Satisfaction (T1)" = c(NA, 
                              correlations_jobsat1[[1,1]],
                              correlations_jobsat1[[2,1]],
                              correlations_jobsat1[[3,1]],
                              correlations_jobsat1[[4,1]],
                              correlations_jobsat1[[5,1]],
                              correlations_jobsat1[[6,1]],
                              correlations_jobsat1[[7,1]],
                              correlations_jobsat1[[8,1]],
                              correlations_jobsat1[[9,1]]),
  "Job Satisfaction (T2)" = c(NA, NA,
                              correlations_jobsat2[[1,1]],
                              correlations_jobsat2[[2,1]],
                              correlations_jobsat2[[3,1]],
                              correlations_jobsat2[[4,1]],
                              correlations_jobsat2[[5,1]],
                              correlations_jobsat2[[6,1]],
                              correlations_jobsat2[[7,1]],
                              correlations_jobsat2[[8,1]]),
  "Belief in Free Will (T1)" = c(NA, NA, NA,
                                 correlations_fw[[1,1]],
                                 correlations_fw[[2,1]],
                                 correlations_fw[[3,1]],
                                 correlations_fw[[4,1]],
                                 correlations_fw[[5,1]],
                                 correlations_fw[[6,1]],
                                 correlations_fw[[7,1]]),

  "Job Autonomy (T1)" = c(NA, NA, NA, NA,
                          correlations_jobaut1[[1,1]],
                          correlations_jobaut1[[2,1]],
                          correlations_jobaut1[[3,1]],
                          correlations_jobaut1[[4,1]],
                          correlations_jobaut1[[5,1]],
                          correlations_jobaut1[[6,1]]),


   "Job Autonomy (T2)" = c(NA, NA, NA, NA, NA,
                           correlations_jobaut1[[1,1]],
                           correlations_jobaut1[[2,1]],
                           correlations_jobaut1[[3,1]],
                           correlations_jobaut1[[4,1]],
                          correlations_jobaut2[[5,1]]),
                          
  check.names = FALSE
)
